@@ -7,7 +7,7 @@
 
 @{
     Layout = "master.cshtml";
-    var author = Model.GetAuthor(Umbraco);
+    var author = Model.Author;
 }
 
 @await Html.PartialAsync("~/Views/Partials/pageHeader.cshtml", new PageHeaderViewModel(Model.Name, Model.Title, Model.Subtitle, Model.MainImage, author.Name, Model.ArticleDate, Model.Categories))
